requirements.txt
langgraph==0.0.69
langchain==0.2.16
langchain-core==0.1.52
langchain-groq==0.1.5
chromadb==0.4.10
streamlit==1.40.0
pydantic==2.10.3
python-dotenv==1.0.1
pypdf==5.1.0
sqlite-utils==3.36
sentence-transformers==2.7.0
fastapi==0.104.1
uvicorn==0.24.0
tabulate==0.9.0
requests==2.31.0
aiofiles==23.2.1